+++ App.js
@@ -1,25 +1,16 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, TextInput, View } from 'react-native';
-import { useState } from 'react';
+import { StyleSheet, Text, View } from 'react-native';
 import Header from './components/Header'
+import Input from './components/Input'
 
 export default function App() {
   const appName = 'Demo App';
-  const [text, setText] = useState("");
-
-  const handleTextChange = (changedText) => {
-    console.log('User is typing.', changedText);
-    setText(changedText);
-  }
 
   return (
     <View style={styles.container}>
       <StatusBar style="auto" />
       <Header name={appName} version={2}/>
-      <TextInput 
-        placeholder='Type something...'
-        value={text}
-        onChangeText={handleTextChange}/> 
+      <Input />
     </View>
     // handleTextChange takes in an onChangeText behind the scenes and we call setText to update the text state
   );
